Untemplate StreamMaster
