+++ src/MasterInterface.h
@@ -8,8 +8,7 @@
 
 namespace FileWriter {
 
-template <typename T> class StreamMaster;
-class Streamer;
+class StreamMaster;
 
 /// \brief Listens to the Kafka configuration topic and handles any requests.
 ///
@@ -36,9 +35,9 @@ public:
   virtual MainOpt &getMainOpt() = 0;
   virtual std::shared_ptr<KafkaW::ProducerTopic> getStatusProducer() = 0;
   virtual void
-  addStreamMaster(std::unique_ptr<StreamMaster<Streamer>> StreamMaster) = 0;
+  addStreamMaster(std::unique_ptr<StreamMaster> StreamMaster) = 0;
   virtual void stopStreamMasters() = 0;
-  virtual std::unique_ptr<StreamMaster<Streamer>> &
+  virtual std::unique_ptr<StreamMaster> &
   getStreamMasterForJobID(std::string const &JobID) = 0;
 };
